# uuid: use cryptographically random ChaCha8 algorithm from math/rand/v2