--- a/pkg/kv/kvserver/gc/gc_test.go
+++ b/pkg/kv/kvserver/gc/gc_test.go
@@ -338,7 +338,7 @@ type testIntent struct {
 func generateScattered(total int, txns int, maxKeySize int, random *rand.Rand) []testIntent {
 	var txnIds []uuid.UUID
 	for len(txnIds) < txns {
-		txnIds = append(txnIds, uuid.FastMakeV4())
+		txnIds = append(txnIds, uuid.MakeV4())
 	}
 	var intents []testIntent
 	for len(intents) < total {
@@ -367,7 +367,7 @@ func generateSequential(total int, maxTxnSize int, maxKeySize int, random *rand.
 	for ; len(intents) < total; leftForTransaction-- {
 		if leftForTransaction == 0 {
 			leftForTransaction = intnFrom1(random, maxTxnSize)
-			txnUUID = uuid.FastMakeV4()
+			txnUUID = uuid.MakeV4()
 		}
 		intents = append(intents,
 			testIntent{
@@ -425,7 +425,7 @@ func TestGCIntentBatcherErrorHandling(t *testing.T) {
 
 	key1 := []byte("key1")
 	key2 := []byte("key2")
-	txn1 := enginepb.MVCCMetadata{Txn: &enginepb.TxnMeta{ID: uuid.FastMakeV4()}}
+	txn1 := enginepb.MVCCMetadata{Txn: &enginepb.TxnMeta{ID: uuid.MakeV4()}}
 
 	// Verify intent cleanup error is propagated to caller.
 	info := Info{}