--- a/pkg/kv/kvserver/loqrecovery/recovery_env_test.go
+++ b/pkg/kv/kvserver/loqrecovery/recovery_env_test.go
@@ -218,10 +218,10 @@ func (*seqUUIDGen) NewV3(uuid.UUID, string) uuid.UUID {
 	panic("not implemented")
 }
 
-func (g *seqUUIDGen) NewV4() (uuid.UUID, error) {
+func (g *seqUUIDGen) NewV4() uuid.UUID {
 	nextV4 := g.next()
 	nextV4.SetVersion(uuid.V4)
-	return nextV4, nil
+	return nextV4
 }
 
 func (*seqUUIDGen) NewV5(uuid.UUID, string) uuid.UUID {