--- a/pkg/server/distsql_flows_test.go
+++ b/pkg/server/distsql_flows_test.go
@@ -31,7 +31,7 @@ func TestMergeDistSQLRemoteFlows(t *testing.T) {
 
 	flowIDs := make([]execinfrapb.FlowID, 4)
 	for i := range flowIDs {
-		flowIDs[i].UUID = uuid.FastMakeV4()
+		flowIDs[i].UUID = uuid.MakeV4()
 	}
 	sort.Slice(flowIDs, func(i, j int) bool {
 		return bytes.Compare(flowIDs[i].GetBytes(), flowIDs[j].GetBytes()) < 0