--- a/pkg/sql/catalog/schematelemetry/schema_telemetry_job.go
+++ b/pkg/sql/catalog/schematelemetry/schema_telemetry_job.go
@@ -88,7 +88,7 @@ func (t schemaTelemetryResumer) Resume(ctx context.Context, execCtx interface{})
 		return err
 	}
 
-	events, err := CollectClusterSchemaForTelemetry(ctx, p.ExecCfg(), asOf, uuid.FastMakeV4(), maxRecords)
+	events, err := CollectClusterSchemaForTelemetry(ctx, p.ExecCfg(), asOf, uuid.MakeV4(), maxRecords)
 	if err != nil || len(events) == 0 {
 		return err
 	}