--- a/pkg/sql/randgen/datum.go
+++ b/pkg/sql/randgen/datum.go
@@ -184,7 +184,7 @@ func RandDatumWithNullChance(
 			sign*rng.Int63n(1000),
 		)}
 	case types.UuidFamily:
-		gen := uuid.NewGenWithReader(rng)
+		gen := uuid.NewGenWithRand(rng.Uint64)
 		return tree.NewDUuid(tree.DUuid{UUID: uuid.Must(gen.NewV4())})
 	case types.INetFamily:
 		ipAddr := ipaddr.RandIPAddr(rng)