--- a/pkg/util/uuid/uuid_wrapper.go
+++ b/pkg/util/uuid/uuid_wrapper.go
@@ -99,16 +99,6 @@ func MakeV4() UUID {
 	return Must(NewV4())
 }
 
-// FastMakeV4 generates a UUID using a fast but not cryptographically secure
-// source of randomness.
-func FastMakeV4() UUID {
-	u, err := fastGen.NewV4()
-	if err != nil {
-		panic(errors.Wrap(err, "should never happen with math/rand.Rand"))
-	}
-	return u
-}
-
 // mathRandReader is an io.Reader that calls through to "math/rand".Read
 // which is safe for concurrent use.
 type mathRandReader struct{}
@@ -119,9 +109,6 @@ func (r mathRandReader) Read(p []byte) (n int, err error) {
 	//lint:ignore SA1019 deprecated
 	return math_rand.Read(p)
 }
-
-// fastGen is a non-cryptographically secure Generator.
-var fastGen = NewGenWithReader(mathRandReader{})
 
 // NewPopulatedUUID returns a populated UUID.
 func NewPopulatedUUID(r interface {