--- a/pkg/workload/movr/workload.go
+++ b/pkg/workload/movr/workload.go
@@ -37,10 +37,7 @@ type movrWorker struct {
 }
 
 func (m *movrWorker) getRandomUser(city string) (string, error) {
-	id, err := uuid.NewV4()
-	if err != nil {
-		return "", err
-	}
+	id := uuid.NewV4()
 	var user string
 	q := `
 		SELECT
@@ -53,15 +50,12 @@ func (m *movrWorker) getRandomUser(city string) (string, error) {
 					(SELECT id FROM users WHERE city = $1 ORDER BY id LIMIT 1) AS b
 			);
 		`
-	err = m.db.QueryRow(q, city, id.String()).Scan(&user)
+	err := m.db.QueryRow(q, city, id.String()).Scan(&user)
 	return user, err
 }
 
 func (m *movrWorker) getRandomPromoCode() (string, error) {
-	id, err := uuid.NewV4()
-	if err != nil {
-		return "", err
-	}
+	id := uuid.NewV4()
 	q := `
 		SELECT
 			IFNULL(a, b)
@@ -74,15 +68,12 @@ func (m *movrWorker) getRandomPromoCode() (string, error) {
 			);
 		`
 	var code string
-	err = m.db.QueryRow(q, id.String()).Scan(&code)
+	err := m.db.QueryRow(q, id.String()).Scan(&code)
 	return code, err
 }
 
 func (m *movrWorker) getRandomVehicle(city string) (string, error) {
-	id, err := uuid.NewV4()
-	if err != nil {
-		return "", err
-	}
+	id := uuid.NewV4()
 	q := `
 		SELECT
 			IFNULL(a, b)
@@ -95,7 +86,7 @@ func (m *movrWorker) getRandomVehicle(city string) (string, error) {
 			);
 		`
 	var vehicle string
-	err = m.db.QueryRow(q, city, id.String()).Scan(&vehicle)
+	err := m.db.QueryRow(q, city, id.String()).Scan(&vehicle)
 	return vehicle, err
 }
 
@@ -263,10 +254,7 @@ func (m *movrWorker) generateWorkSimulation() func(context.Context) error {
 
 	return func(ctx context.Context) error {
 		activeCity := randCity(m.rng)
-		id, err := uuid.NewV4()
-		if err != nil {
-			return err
-		}
+		id := uuid.NewV4()
 		// Our workload is as follows: with 95% chance, do a simple read operation.
 		// Else, update all active vehicle locations, then pick a random "write" operation
 		// weighted by the weights in movrWorkloadFns.
@@ -275,7 +263,7 @@ func (m *movrWorker) generateWorkSimulation() func(context.Context) error {
 				return m.readVehicles(activeCity)
 			})
 		}
-		err = runAndRecord("updateActiveRides", func() error {
+		err := runAndRecord("updateActiveRides", func() error {
 			return m.updateActiveRides()
 		})
 		if err != nil {