fixes https://github.com/cockroachdb/cockroach/issues/109397

###  uuid: add parallel benchmark

This lets us test the effect of sharing a global random source
between multiple threads.

### uuid: use mathRandReader instead of crypto/rand

go1.22 uses a new random number generator called ChaCha8, and it is
cryptographically random. We can use this instead of the slower
crypto/rand random source to ensure UUID uniqueness.

https://go.dev/blog/chacha8rand

### *: rename FastMakeV4 to MakeV4

There's no need for a separate FastMakeV4 function now; it's the same as
MakeV4.

###  uuid: use math/rand/v2

This new rand package is an improvement since it makes it impossible to
accidentally set a global seed for the shared RNG, which would make
UUIDs become entirely predictable and collision-prone.

It also appears to be faster. I think it's because now the global RNG
does not need to be guarded by a mutex.

```
Removing the top-level Seed also let us hard-code the use of scalable,
per-thread generators by the top-level methods, avoiding a GODEBUG check
at each use.
```

See:
- https://go.dev/blog/randv2
- golang/go@0b9974d

### uuid: remove impossible error cases; reduce branching

This consolidates some conditional branches so that the Uint64()
function is always used to generate random numbers rather than using
(io.Reader).Read(). This change makes it so some functions can no longer
return errors, so some tests were removed.

---

### Diff between `crypto/rand` and `math/rand/v1`
```
goos: darwin
goarch: arm64
                       │ bench-crypto-rand.txt │    bench-math-rand-v1-trial1.txt    │
                       │        sec/op         │   sec/op     vs base                │
FastMakeV4-8                       29.49n ± 1%   29.68n ± 1%   +0.63% (p=0.011 n=10)
MakeV4-8                          519.50n ± 0%   29.52n ± 0%  -94.32% (p=0.000 n=10)
ConcurrentFastMakeV4-8             133.2n ± 0%   133.0n ± 0%   -0.19% (p=0.048 n=10)
ConcurrentMakeV4-8                 464.6n ± 1%   133.0n ± 0%  -71.36% (p=0.000 n=10)
geomean                            175.5n        62.75n       -64.24%

                       │ bench-crypto-rand.txt │      bench-math-rand-v1-trial1.txt      │
                       │         B/op          │    B/op     vs base                     │
FastMakeV4-8                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
MakeV4-8                          16.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
ConcurrentFastMakeV4-8            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ConcurrentMakeV4-8                16.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
geomean                                      ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                       │ bench-crypto-rand.txt │      bench-math-rand-v1-trial1.txt      │
                       │       allocs/op       │ allocs/op   vs base                     │
FastMakeV4-8                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
MakeV4-8                          1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
ConcurrentFastMakeV4-8            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ConcurrentMakeV4-8                1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                                      ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```

### Diff between `math/rand/v1` and `math/rand/v2`
```
goos: darwin
goarch: arm64
                   │ bench-math-rand-v1-trial2.txt │       bench-math-rand-v2.txt        │
                   │            sec/op             │   sec/op     vs base                │
MakeV4-8                               29.44n ± 1%   12.18n ± 1%  -58.61% (p=0.000 n=10)
ConcurrentMakeV4-8                   133.100n ± 0%   1.642n ± 2%  -98.77% (p=0.000 n=10)
geomean                                62.60n        4.472n       -92.86%

                   │ bench-math-rand-v1-trial2.txt │       bench-math-rand-v2.txt        │
                   │             B/op              │    B/op     vs base                 │
MakeV4-8                              0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ConcurrentMakeV4-8                    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                          ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                   │ bench-math-rand-v1-trial2.txt │       bench-math-rand-v2.txt        │
                   │           allocs/op           │ allocs/op   vs base                 │
MakeV4-8                              0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
ConcurrentMakeV4-8                    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                          ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

### Overall improvement from this PR
```
goos: darwin
goarch: arm64
                       │ bench-crypto-rand.txt │        bench-math-rand-v2.txt         │
                       │        sec/op         │   sec/op     vs base                  │
FastMakeV4-8                       29.49n ± 1%
MakeV4-8                          519.50n ± 0%   12.18n ± 1%  -97.65% (p=0.000 n=10)
ConcurrentFastMakeV4-8             133.2n ± 0%
ConcurrentMakeV4-8               464.600n ± 1%   1.642n ± 2%  -99.65% (p=0.000 n=10)
geomean                            175.5n        4.472n       -99.09%                ¹
¹ benchmark set differs from baseline; geomeans may not be comparable

                       │ bench-crypto-rand.txt │          bench-math-rand-v2.txt          │
                       │         B/op          │   B/op     vs base                       │
FastMakeV4-8                      0.000 ± 0%
MakeV4-8                          16.00 ± 0%     0.00 ± 0%  -100.00% (p=0.000 n=10)
ConcurrentFastMakeV4-8            0.000 ± 0%
ConcurrentMakeV4-8                16.00 ± 0%     0.00 ± 0%  -100.00% (p=0.000 n=10)
geomean                                      ¹              ?                       ² ¹ ³
¹ summaries must be >0 to compute geomean
² benchmark set differs from baseline; geomeans may not be comparable
³ ratios must be >0 to compute geomean

                       │ bench-crypto-rand.txt │          bench-math-rand-v2.txt           │
                       │       allocs/op       │ allocs/op   vs base                       │
FastMakeV4-8                      0.000 ± 0%
MakeV4-8                          1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
ConcurrentFastMakeV4-8            0.000 ± 0%
ConcurrentMakeV4-8                1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                                      ¹               ?                       ² ¹ ³
¹ summaries must be >0 to compute geomean
² benchmark set differs from baseline; geomeans may not be comparable
³ ratios must be >0 to compute geomean
```